rename information to conceptual submodul
