+++ cognite/neat/core/_data_model/_shared.py
@@ -5,10 +5,10 @@ from cognite.neat.core._data_model.models import (
     DMSRules,
     InformationRules,
 )
-from cognite.neat.core._data_model.models.dms._rules_input import DMSInputRules
-from cognite.neat.core._data_model.models.information._rules_input import (
+from cognite.neat.core._data_model.models.conceptual._rules_input import (
     InformationInputRules,
 )
+from cognite.neat.core._data_model.models.dms._rules_input import DMSInputRules
 from cognite.neat.core._utils.spreadsheet import SpreadsheetRead
 
 VerifiedRules: TypeAlias = InformationRules | DMSRules
