+++ cognite/neat/core/_data_model/models/conceptual/_rules_input.py
