+++ cognite/neat/core/_data_model/transformers/_verification.py
@@ -13,8 +13,8 @@ from cognite.neat.core._data_model.models import (
     InformationInputRules,
     InformationRules,
 )
+from cognite.neat.core._data_model.models.conceptual import InformationValidation
 from cognite.neat.core._data_model.models.dms import DMSValidation
-from cognite.neat.core._data_model.models.information import InformationValidation
 from cognite.neat.core._issues import MultiValueError, catch_issues
 from cognite.neat.core._issues.errors import NeatTypeError, NeatValueError
